+++ tests/e2e-tests/src/main/kotlin/api_models/Credential.kt
@@ -11,6 +11,8 @@ data class Credential(
     var subjectId: String = "",
     var updatedAt: String = "",
     var validityPeriod: Int = 0,
-    var claims: LinkedHashMap<String,String> = LinkedHashMap(),
-    var jwtCredential: String = ""
+    var claims: LinkedHashMap<String, String> = LinkedHashMap(),
+    var jwtCredential: String = "",
+    var issuingDID: String = "",
+    var connectionId: String = "",
 )