+++ tests/e2e-tests/src/main/kotlin/api_models/PublicKey.kt
@@ -4,7 +4,7 @@ import com.fasterxml.jackson.annotation.JsonValue
 
 data class PublicKey(
     val id: String,
-    val purpose: Purpose
+    val purpose: Purpose,
 )
 
 enum class Purpose(@JsonValue val value: String) {
@@ -12,5 +12,5 @@ enum class Purpose(@JsonValue val value: String) {
     ASSERTION_METHOD("assertionMethod"),
     KEY_AGREEMENT("keyAgreement"),
     CAPABILITY_INVOCATION("capabilityInvocation"),
-    CAPABILITY_DELEGATION("capabilityDelegation")
+    CAPABILITY_DELEGATION("capabilityDelegation"),
 }